+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - REST / Add Request.profile(Class) typed user profile getter
 - REST / Add Xtream.scan() processing method
+- REST / Expose JSONLDFOrmat.encode/decode() utility methods
 - JSON / Extend JSONPath.Processor API with getters for JSON objects, arrays and nested processors
 - XML / Add Untag.untag(String) utility method 
 - RDF / Factor RDFFormat static parsing method